Reconfigured project and cleaned errors
Reconfigured project to be source file. Cleaned up most errors except for two remaining in Shelf.java and Multimap,java.

Errors in Multimap are with add and remove method. Passing 1 key and 2 values instead of 1 key and 1 value.

Errors with Shelf.java include return of Collection method and declaration of variables.